fix ImportAsStocktakeJob

+++ app/Modules/DataCollector/src/Jobs/ImportAsStocktakeJob.php
@@ -6,7 +6,6 @@ use App\Abstracts\UniqueJob;
 use App\Models\DataCollection;
 use App\Models\DataCollectionRecord;
 use App\Models\DataCollectionStocktake;
-use App\Models\Inventory;
 use App\Models\InventoryMovement;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -28,34 +27,26 @@ class ImportAsStocktakeJob extends UniqueJob
 
     public function handle(): void
     {
-        /** @var DataCollection $sourceDataCollection */
+        /** @var DataCollection $dataCollection */
         $dataCollection = DataCollection::withTrashed()->findOrFail($this->dataCollection_id);
 
         $inventoryMovementRecords = $dataCollection->records()
-            ->where('quantity_scanned', '!=', DB::raw(0))
             ->get()
             ->map(function (DataCollectionRecord $record) use ($dataCollection) {
-                /** @var Inventory $inventory */
-                $inventory = Inventory::where([
-                    'product_id' => $record->product_id,
-                    'warehouse_id' => $dataCollection->warehouse_id,
-                ])
-                    ->first();
-
-                $quantityDelta = $record->quantity_scanned - $inventory->quantity;
-
                 $custom_uuid = implode('-', ['source_data_collections_records_id', $record->getKey()]);
 
                 return [
                     'custom_unique_reference_id' => $custom_uuid,
-                    'occurred_at' => now()->utc()->toDateTimeLocalString(),
-                    'inventory_id' => $inventory->id,
+                    'sequence_number' => null,
+                    'occurred_at' => $dataCollection->deleted_at ?? now()->utc()->toDateTimeLocalString(),
+                    'inventory_id' => $record->inventory_id,
                     'type' => InventoryMovement::TYPE_STOCKTAKE,
-                    'product_id' => $inventory->product_id,
-                    'warehouse_id' => $inventory->warehouse_id,
-                    'quantity_before' => $inventory->quantity,
-                    'quantity_delta' => $quantityDelta,
-                    'quantity_after' => $inventory->quantity + $quantityDelta,
+                    'product_id' => $record->product_id,
+                    'warehouse_id' => $dataCollection->warehouse->id,
+                    'warehouse_code' => $dataCollection->warehouse->code,
+                    'quantity_before' => $record->inventory->quantity,
+                    'quantity_delta' => $record->quantity_scanned - $record->inventory->quantity,
+                    'quantity_after' => $record->quantity_scanned,
                     'description' => Str::substr('Data Collection - ' . $dataCollection->name, 0, 50),
                     'user_id' => Auth::id(),
                     'created_at' => now()->utc()->toDateTimeLocalString(),
@@ -64,11 +55,13 @@ class ImportAsStocktakeJob extends UniqueJob
             });
 
         DB::transaction(function () use ($dataCollection, $inventoryMovementRecords) {
-            $dataCollection->update(['type' => DataCollectionStocktake::class]);
+            InventoryMovement::query()->upsert($inventoryMovementRecords->toArray(), ['custom_unique_reference_id'], ['quantity_delta', 'quantity_after', 'updated_at']);
 
-            $dataCollection->delete();
+            if ($dataCollection->deleted_at === null) {
+                $dataCollection->delete();
+            }
 
-            InventoryMovement::query()->insert($inventoryMovementRecords->toArray());
+            $dataCollection->update(['type' => DataCollectionStocktake::class, 'currently_running_task' => null]);
         });
     }
 }
